+++ Package.swift
@@ -10,7 +10,7 @@ let package = Package(
   dependencies: [
     .package(url: "https://github.com/galileostd/swift-msgpack", from: "1.3.0"),
     .package(url: "https://github.com/apple/swift-crypto", from: "4.5.0"),
-    .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.5.0"),
+    .package(url: "https://github.com/apple/swift-log", from: "1.14.0"),
   ],
   targets: [
     .target(
@@ -18,6 +18,7 @@ let package = Package(
       dependencies: [
         .product(name: "SwiftMsgpack", package: "swift-msgpack"),
         .product(name: "Crypto", package: "swift-crypto"),
+        .product(name: "Logging", package: "swift-log"),
       ],
       path: "Sources/NyaruDB2",
       linkerSettings: [